Fix readme indendation and gofmt
I'm actually not sure what the "bt" package is supposed to be since it's
not part of the stdlib and it doesn't match the current godoc. Therefore
I'm assuming it's a project-local import and putting it at the bottom.
This is probably a sign of more explanation needed (or at least a
fully-qualified import path and explaining that it sets up your Tracer
for you).